+++ tests/testthat/test-FPearson.R
@@ -1,18 +1,18 @@
-context("FPearson functions")
+context("Fcorr functions")
 
-# We check the functions pFPearson, qFPearson and rFPearson.
+# We check the functions pFcorr, qFcorr and rFcorr.
 
 # Set a tolerance for the comparison of the simulated values
 my_tol <- 1e-5
 
-# 1. Check that calling qFPearson with probabilities p and then calling pgev
+# 1. Check that calling qFcorr with probabilities p and then calling pgev
 #    with the results gets us back to the initial probabilities.
 
-pqFPearson_test_fn <- function(x, p) {
+pqFcorr_test_fn <- function(x, p) {
   N <- x[1]
   rho <- x[2]
-  qs <- qFPearson(p = p, N = N, rho = rho)
-  ps <- pFPearson(qs, N = N, rho = rho)
+  qs <- qFcorr(p = p, N = N, rho = rho)
+  ps <- pFcorr(qs, N = N, rho = rho)
   return(list(p = p, ps = ps))
 }
 
@@ -28,22 +28,22 @@ rho_check <- c(-0.99, -0.5, -0.1, -ep, 0, ep, 0.1, 0.5, 0.99)
 par_vals <- cbind(N_check, rho_check)
 p_vals <- c(0.01, 0.1, 0.5, 0.9, 0.99)
 for (i in 1:nrow(par_vals)) {
-  test_string <- paste("p and q, FPearson (rho, N) = ", par_vals[i, ])
-  x <- pqFPearson_test_fn(x = par_vals[i, ], p = p_vals)
+  test_string <- paste("p and q, Fcorr (rho, N) = ", par_vals[i, ])
+  x <- pqFcorr_test_fn(x = par_vals[i, ], p = p_vals)
   test_function(x, test_string)
 }
 
-# 2. Similar to 1. but now generate the quantiles using rFPearson().
+# 2. Similar to 1. but now generate the quantiles using rFcorr().
 
 seed <- 19022018
 set.seed(seed)
 
-rpqFPearson_test_fn <- function(x, p) {
+rpqFcorr_test_fn <- function(x, p) {
   N <- x[1]
   rho <- x[2]
-  qs_in <- rFPearson(n = 5, N = N, rho = rho)
-  ps <- pFPearson(qs, N = N, rho = rho)
-  qs_out <- qFPearson(p = p, N = N, rho = rho)
+  qs_in <- rFcorr(n = 5, N = N, rho = rho)
+  ps <- pFcorr(qs, N = N, rho = rho)
+  qs_out <- qFcorr(p = p, N = N, rho = rho)
   return(list(qs_in = qs_in, qs_out = qs_out))
 }
 
@@ -59,8 +59,8 @@ rho_check <- c(-0.99, -0.5, -0.1, -ep, 0, ep, 0.1, 0.5, 0.99)
 par_vals <- cbind(N_check, rho_check)
 p_vals <- c(0.01, 0.1, 0.5, 0.9, 0.99)
 for (i in 1:nrow(par_vals)) {
-  test_string <- paste("p, q and r, FPearson (rho, N) = ", par_vals[i, ])
-  x <- pqFPearson_test_fn(x = par_vals[i, ], p = p_vals)
+  test_string <- paste("p, q and r, Fcorr (rho, N) = ", par_vals[i, ])
+  x <- pqFcorr_test_fn(x = par_vals[i, ], p = p_vals)
   test_function(x, test_string)
 }