+++ NEWS.md
@@ -1,5 +1,9 @@
 # smovie 1.1.2.9000
 
+## UCL Eugenics Inquiry
+
+Explicit references to (Karl) Pearson have removed from the package following the [Inquiry into the History of Eugenics at UCL](https://www.ucl.ac.uk/provost/inquiry-history-eugenics-ucl).
+
 ## Bug fixes and minor improvements
 
 * In `wws()` the gradient of the log-likelihood was unnecessarily labelled with a red "score" when the null value of theta0 is less than the MLE of theta.  This has been removed.
@@ -18,7 +22,7 @@
 
 * New movie: `mean_vs_median()`.  Compares the sampling distributions of the mean and median for random samples from either a standard normal distribution or a standard Student t distribution.
 
-* pkgdown documentation at https://paulnorthrop.github.io/smovie/
+* pkgdown documentation at [https://paulnorthrop.github.io/smovie/](https://paulnorthrop.github.io/smovie/).
 
 ## Bug fixes and minor improvements